requirements.txt
torch
torchvision
tqdm
PyYAML
scikit-learn
scikit-image
opencv-python
pandas
numpy
joblib
Pillow
matplotlib
